chore: rename "-types.ts" file
