+++ packages/core/src/features/index.ts
@@ -1,1 +1,1 @@
-export * from './-types';
+export * from './types';
